エラー起きるので直した

+++ main.rb
@@ -56,20 +56,24 @@ puts startStr + "から" + endStr + "までのデータを取得します"
 # 処理部(DLして保存)
 # レポート
 (Date.parse(timeStart)..Date.parse(timeEnd)).each do |date|
-    begin
-        ymd = date.strftime("%Y%m%d")
-        url = "https://so2-api.mutoys.com/json/report/buy#{ymd}.json"
-        puts url + "にアクセスしています"
-        open(url) do |rep|
-            open("./so2stockdata/buy_reports/#{ymd}.json", "w+b") do |save|
-                save.write(rep.read)
+    ymd = date.strftime("%Y%m%d")
+    if File.exist?("./so2stockdata/buy_reports/#{ymd}.json") == false
+        begin
+            url = "https://so2-api.mutoys.com/json/report/buy#{ymd}.json"
+            puts url + "にアクセスしています"
+            open(url) do |rep|
+                open("./so2stockdata/buy_reports/#{ymd}.json", "w+b") do |save|
+                    save.write(rep.read)
+                end
             end
+            puts "./so2stockdata/buy_reports/#{ymd}.jsonとして保存しました"
+        rescue
+            puts "エラー: " + url + "にアクセスできませんでした"
         end
-        puts "./so2stockdata/buy_reports/#{ymd}.jsonとして保存しました"
-    rescue
-        puts "エラー: " + url + "にアクセスできませんでした"
+        sleep(0.4)
+    else
+        puts "./so2stockdata/buy_reports/#{ymd}.jsonは既に存在しています"
     end
-    sleep(0.3)
 end
 
 # 月間全部門トップ3
@@ -88,15 +92,17 @@ end
         rescue
             puts "エラー: " + url + "にアクセスできませんでした"
         end
-        sleep(0.3)
+        sleep(0.4)
+    else
+        puts "./so2stockdata/ranking/all_top3_monthly/#{ymd}.jsonは既に存在しています"
     end
 end
 
 # 月間部門別トップ1000
 (Date.parse(timeStart)..Date.parse(timeEnd)).each do |date|
     ymd = date.strftime("%Y-%m")
-    if File.exist?("./so2stockdata/ranking/top1000_monthly/#{departmentArray[element]}/#{ymd}.json") == false
-        (departmentArray.length).times do |element|
+    (departmentArray.length).times do |element|
+        if File.exist?("./so2stockdata/ranking/top1000_monthly/#{departmentArray[element]}/#{ymd}.json") == false
             begin
                 url = "https://so2-api.mutoys.com/json/ranking/#{ymd}/#{departmentArray[element]}.json"
                 puts url + "にアクセスしています"
@@ -109,28 +115,34 @@ end
             rescue
                 puts "エラー: " + url + "にアクセスできませんでした"
             end
-            sleep(0.3)
+            sleep(0.4)
+        else
+            puts "./so2stockdata/ranking/top1000_monthly/#{departmentArray[element]}/#{ymd}.jsonは既に存在しています"
         end
-        sleep(0.3)
     end
+    sleep(0.4)
 end
 
 # デイリートップ1000(部門別)
 (Date.parse(timeStart)..Date.parse(timeEnd)).each do |date|
     ymd = date.strftime("%Y-%m-%d")
     (departmentArray.length).times do |element|
-        begin
-            url = "https://so2-api.mutoys.com/json/ranking/#{ymd}/#{departmentArray[element]}.json"
-            puts url + "にアクセスしています"
-            open(url) do |rep|
-                open("./so2stockdata/ranking/top1000_daily/#{departmentArray[element]}/#{ymd}.json", "w+b") do |save|
-                    save.write(rep.read)
-               end
+        if File.exist?("./so2stockdata/ranking/top1000_daily/#{departmentArray[element]}/#{ymd}.json")
+            begin
+                url = "https://so2-api.mutoys.com/json/ranking/#{ymd}/#{departmentArray[element]}.json"
+                puts url + "にアクセスしています"
+                open(url) do |rep|
+                    open("./so2stockdata/ranking/top1000_daily/#{departmentArray[element]}/#{ymd}.json", "w+b") do |save|
+                        save.write(rep.read)
+                   end
+                end
+                puts "./so2stockdata/ranking/top1000_daily/#{departmentArray[element]}/#{ymd}.jsonとして保存しました"
+            rescue
+                puts "エラー: " + url + "にアクセスできませんでした"
             end
-            puts "./so2stockdata/ranking/top1000_daily/#{departmentArray[element]}/#{ymd}.jsonとして保存しました"
-        rescue
-            puts "エラー: " + url + "にアクセスできませんでした"
+            sleep(0.4)
+        else
+            puts "./so2stockdata/ranking/top1000_daily/#{departmentArray[element]}/#{ymd}.jsonは既に存在しています"
         end
-        sleep(0.3)
     end
 end
